v2.2.7
Previously, the map for events in kal3000 was shown, as soon as "Angezeigte Adresse (ohne Stadt) was present and not empty. But for the map to really work, the coordinates have had to been set. So now we make sure, the map's only shown if this is the case.

Also the process of setting the coordinates was quite hard to understand, so this update provides a little helper text on this.