Allow text rendering as early as possible
By setting font-display we allow the text to be rendered even if the font is
not yet loaded. For headlines we swap the font - even if it is loaded very
late. for text the user might already have started reading and it might confuse
him if the font changes, so we leave it be if the font takes to long.

See https://developers.google.com/web/updates/2016/02/font-display